add redux-thunk for DalToBll and refactor Auth
